local storage works. bugfixes. cleanup. yay!!!

+++ src/app/services/colors.service.ts
@@ -1,47 +1,83 @@
 import { Injectable } from '@angular/core';
 import { collection, collectionData, doc, Firestore, query, setDoc, Timestamp, where } from '@angular/fire/firestore';
 import { deleteDoc } from '@firebase/firestore';
-import { Observable } from 'rxjs';
+import { BehaviorSubject, Observable, Subscriber } from 'rxjs';
 import { Color } from '../interfaces/color';
+import { AuthService } from './auth.service';
 
 @Injectable({
   providedIn: 'root'
 })
 export class ColorsService {
-  private colors$ = new Observable<Color[]>();
+  private colors$ = new BehaviorSubject<Color[]>([]);
+  private colors: Color[] = [];
 
-  constructor(private firestore: Firestore) {
+  constructor(
+    private firestore: Firestore,
+    private authService: AuthService) {
   }
 
-  async load(userId: string | null) {
-    if (userId !== null) {
+  async load(userId: string | null | undefined) {
+    if (userId) {
+      // if there's a user ID, that means we can try to pull from Firestore
       const colorCollection = collection(this.firestore, 'colors');
       const queriedData = query(colorCollection, where("userId", "==", userId));
-      this.colors$ = collectionData(queriedData) as Observable<Color[]>;
+      this.colors$ = collectionData(queriedData) as BehaviorSubject<Color[]>;
     } else {
-      this.colors$ = new Observable<Color[]>();
+      // there's no user ID; pull from local storage
+      const locallyStoredColors = localStorage.getItem('colors');
+      if (locallyStoredColors !== null) {
+        // if there are locally stored colors, then load those
+        this.colors = JSON.parse(locallyStoredColors);
+      }
+      this.colors$ = new BehaviorSubject<Color[]>(this.colors);
     }
   }
 
-  getColors(): Observable<Color[]> {
+  getColors(): BehaviorSubject<Color[]> {
     return this.colors$;
   }
 
   async removeColor(colorId: string): Promise<void> {
-    await deleteDoc(doc(this.firestore, 'colors', colorId));
+    const currentUser = this.authService.getSignedInUid();
+    if (currentUser) {
+      // we're logged in; delete from firestore
+      await deleteDoc(doc(this.firestore, 'colors', colorId));
+    } else {
+      // not logged in; delete locally
+      const index = this.colors.findIndex((color: Color) => color.id === colorId);
+      this.colors.splice(index, 1);
+      localStorage.setItem('colors', JSON.stringify(this.colors));
+    }
   }
 
-  async addColor(userId: string, value: string, name?: string, description?: string): Promise<string> {
+  async addColor(value: string, name: string, description?: string): Promise<string> {
     const docRef = doc(collection(this.firestore, 'colors'));
 
-    await setDoc(docRef, {
-      id: docRef.id,
-      userId,
-      value,
-      name,
-      description,
-      time: Timestamp.now(),
-    });
+    const currentUser = this.authService.getSignedInUid();
+    if (currentUser) {
+      // online storage
+      await setDoc(docRef, {
+        id: docRef.id,
+        userId: currentUser,
+        value,
+        name,
+        description,
+        time: Timestamp.now(),
+      });
+    } else {
+      // offline storage
+      this.colors.push({
+        id: docRef.id,
+        userId: '',
+        name,
+        description,
+        value,
+        time: Timestamp.now(),
+      });
+      this.colors$.next(this.colors);
+      localStorage.setItem('colors', JSON.stringify(this.colors));
+    }
 
     return docRef.id;
   }
